+If merge non-equal composition error
+Correctly throws errors when merging composed & decomposed states
